Renaming app folder
